backend/requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
